==> 2.0.convert_to_prata.ipynb <==
import pandas as pd
from pyspark.sql import SparkSession,Window
import pyspark.sql.functions as F
import seaborn as sns
import os
import duckdb
from pathlib import Path

spark = SparkSession.builder.appName("Mateus") \
    .config("spark.sql.shuffle.partitions", "100") \
    .config("spark.sql.files.maxPartitionBytes", "128MB") \
    .config("spark.sql.files.openCostInBytes", "4MB")  \
    .config("spark.driver.maxResultSize", "4g") \
    .config("spark.driver.memory", "16g")  \
    .config("spark.executor.cores", "4") \
    .config("spark.executor.memory", "8g") \
    .config("spark.sql.execution.arrow.enabled", "true") \
    .config("spark.sql.execution.arrow.pyspark.enabled", "true") \
    .config("spark.dynamicAllocation.enabled", "true") \
    .config("spark.dynamicAllocation.minExecutors", "1") \
    .config("spark.dynamicAllocation.maxExecutors", "4")  \
    .getOrCreate()

with open('initial_path.txt','r') as f:
    initial_path = f.read()
    
bronze_path = os.path.join(initial_path,'1. bronze')
prata_path = os.path.join(initial_path,'2. prata')
conn = duckdb.connect('r2_prata.duckdb')

def compute_lat_delta(df):
    lat_max = df.select(F.max('lat')).collect()[0][0]
    lat_min = df.select(F.min('lat')).collect()[0][0]
    lat_range = lat_max-lat_min
    lat_n = df.select('lat').distinct().count()
    lat_delta = lat_range/(lat_n-1)
    return lat_delta

def compute_lon_delta(df):
    lon_max = df.select(F.max('lon')).collect()[0][0]
    lon_min = df.select(F.min('lon')).collect()[0][0]
    lon_range = lon_max-lon_min
    lon_n = df.select('lon').distinct().count()
    lon_delta = lon_range/(lon_n-1)
    return lon_delta

# 1. Separando tabelas dim e tabelas fato

## dim_estacoes

grupo_estacao = ['ana','cemaden','cemig','iac','inmet_agri','inmet','unesp']

estacao = {
    nome.upper():
    pd.read_csv(os.path.join(bronze_path,f'{nome}.csv')) 
    for nome in grupo_estacao
}

id_estacao = {
    nome.upper():
    list(df['id_estacao'].unique())
    for nome,df in estacao.items()
}

dim_estacoes_dict = {}
for nome,df in estacao.items():
    dim_df = df \
        .groupby(['id_estacao','lat','lon'],as_index=False).agg(
            dt_min = ('data','min'),
            dt_max = ('data','max'),
            fl_tmin = ('tmin','max'),
            fl_tmed = ('tmed','max'),
            fl_tmax = ('tmax','max'),
            fl_prec = ('prec','max'),
            fl_urmin = ('urmin','max'),
            fl_urmed = ('urmed','max'),
            fl_urmax = ('urmax','max')
            )
    for col in [c for c in dim_df.columns if c.startswith('fl_')]:
        dim_df[col] = dim_df[col].notnull().astype(int)
    dim_df['nm_grupo_estacao'] = nome.upper()
    dim_estacoes_dict[nome] = dim_df

dim_estacoes  = pd.concat(dim_estacoes_dict.values()).reset_index(drop=True)
conn.execute("CREATE TABLE IF NOT EXISTS dim_estacoes AS SELECT * FROM dim_estacoes")

conn.execute("SHOW TABLES").fetchall()

## fato_satelites

# Renomeando as colunas
satelites_nomes = ['AgCFSR','AgMERRA','CHIRPS','CPC','GL','GPM Final Run','GPM Late Run','PERSIANN-CDR','POWER','TRMM']
satelite_dict = {nome:spark.read.csv(os.path.join(bronze_path,f'full_{nome}.csv'),header=True,inferSchema=True) for nome in satelites_nomes}

satelite_dict['AgCFSR'] = satelite_dict['AgCFSR'] \
    .withColumnRenamed('data','dt_medicao') \
    .withColumnRenamed('vl_prate','vl_precipitacao') \
    .withColumnRenamed('vl_rhstmax','vl_umidade_relativa_maxima') \
    .withColumnRenamed('vl_srad','vl_radiacao_solar') \
    .withColumnRenamed('vl_tavg','vl_temperatura_media') \
    .withColumnRenamed('vl_tmax','vl_temperatura_maxima') \
    .withColumnRenamed('vl_tmin','vl_temperatura_minima') \
    .withColumnRenamed('vl_wndspd','vl_vento_velocidade')

satelite_dict['AgMERRA'] = satelite_dict['AgMERRA'] \
    .withColumnRenamed('data','dt_medicao') \
    .withColumnRenamed('vl_prate','vl_precipitacao') \
    .withColumnRenamed('vl_rhstmax','vl_umidade_relativa_maxima') \
    .withColumnRenamed('vl_srad','vl_radiacao_solar') \
    .withColumnRenamed('vl_tavg','vl_temperatura_media') \
    .withColumnRenamed('vl_tmax','vl_temperatura_maxima') \
    .withColumnRenamed('vl_tmin','vl_temperatura_minima') \
    .withColumnRenamed('vl_wndspd','vl_vento_velocidade')

satelite_dict['CHIRPS'] = satelite_dict['CHIRPS'] \
    .withColumnRenamed('data','dt_medicao') \
    .withColumnRenamed('vl_precipitation','vl_precipitacao')

satelite_dict['CPC'] = satelite_dict['CPC'] \
    .withColumnRenamed('data','dt_medicao') \
    .withColumnRenamed('vl_precipitation','vl_precipitacao') \
    .withColumnRenamed('vl_tmax','vl_temperatura_maxima') \
    .withColumnRenamed('vl_tmin','vl_temperatura_minima')

satelite_dict['GL'] = satelite_dict['GL'] \
    .withColumnRenamed('data','dt_medicao')

satelite_dict['GPM Final Run'] = satelite_dict['GPM Final Run'] \
    .withColumnRenamed('data','dt_medicao') \
    .withColumnRenamed('vl_precipitation','vl_precipitacao')

satelite_dict['GPM Late Run'] = satelite_dict['GPM Late Run'] \
    .withColumnRenamed('data','dt_medicao') \
    .withColumnRenamed('vl_precipitation','vl_precipitacao')

satelite_dict['PERSIANN-CDR'] = satelite_dict['PERSIANN-CDR'] \
    .withColumnRenamed('data','dt_medicao') \
    .withColumnRenamed('vl_precipitation','vl_precipitacao')

satelite_dict['POWER'] = satelite_dict['POWER'] \
    .withColumnRenamed('data','dt_medicao') \
    .withColumnRenamed('vl_allsky_sfc_sw_dwn','vl_radiacao_solar_ceu_total') \
    .withColumnRenamed('vl_prectotcorr','vl_precipitacao_total_corrigido') \
    .withColumnRenamed('vl_ps','vl_pressao_superficie') \
    .withColumnRenamed('vl_rh2m','vl_umidade_relativa_2m') \
    .withColumnRenamed('vl_t2m_max','vl_temperatura_maxima_2m') \
    .withColumnRenamed('vl_t2m_min','vl_temperatura_minima_2m') \
    .withColumnRenamed('vl_t2m','vl_temperatura_media_2m') \
    .withColumnRenamed('vl_t2mdew','vl_temperatura_orvalho_2m') \
    .withColumnRenamed('vl_wd2m','vl_direcao_vento_2m') \
    .withColumnRenamed('vl_wd10m','vl_direcao_vento_10m') \
    .withColumnRenamed('vl_ws2m_max','vl_maxima_vento_2m') \
    .withColumnRenamed('vl_ws10m_max','vl_maxima_vento_10m') \
    .withColumnRenamed('vl_ws10m','vl_vento_10m')

satelite_dict['TRMM'] = satelite_dict['TRMM'] \
    .withColumnRenamed('data','dt_medicao') \
    .withColumnRenamed('vl_precipitation','vl_precipitacao')

print('AgCFSR')
satelite_dict['AgCFSR'] \
        .coalesce(1) \
        .write.option("header",True) \
        .option("delimiter",",") \
        .mode("overwrite") \
        .csv(os.path.join(prata_path, f'prata_fato_AgCFSR.csv'))

print('AgMERRA')
satelite_dict['AgMERRA'] \
        .coalesce(1) \
        .write.option("header",True) \
        .option("delimiter",",") \
        .mode("overwrite") \
        .csv(os.path.join(prata_path, f'prata_fato_AgMERRA.csv'))

print('CHIRPS')
satelite_dict['CHIRPS'] \
        .coalesce(1) \
        .write.option("header",True) \
        .option("delimiter",",") \
        .mode("overwrite") \
        .csv(os.path.join(prata_path, f'prata_fato_CHIRPS.csv'))

print('CPC')
satelite_dict['CPC'] \
        .coalesce(1) \
        .write.option("header",True) \
        .option("delimiter",",") \
        .mode("overwrite") \
        .csv(os.path.join(prata_path, f'prata_fato_CPC.csv'))

print('GL')
satelite_dict['GL'] \
        .coalesce(1) \
        .write.option("header",True) \
        .option("delimiter",",") \
        .mode("overwrite") \
        .csv(os.path.join(prata_path, f'prata_fato_GL.csv'))

print('GPM Final Run')
satelite_dict['GPM Final Run'] \
        .coalesce(1) \
        .write.option("header",True) \
        .option("delimiter",",") \
        .mode("overwrite") \
        .csv(os.path.join(prata_path, f'prata_fato_GPM_Final_Run.csv'))

print('GPM Late Run')
satelite_dict['GPM Late Run'] \
        .coalesce(1) \
        .write.option("header",True) \
        .option("delimiter",",") \
        .mode("overwrite") \
        .csv(os.path.join(prata_path, f'prata_fato_GPM_Late_Run.csv'))

print('PERSIANN-CDR')
satelite_dict['PERSIANN-CDR'] \
        .coalesce(1) \
        .write.option("header",True) \
        .option("delimiter",",") \
        .mode("overwrite") \
        .csv(os.path.join(prata_path, f'prata_fato_PERSIANN_CDR.csv'))

print('TRMM')
satelite_dict['TRMM'] \
        .coalesce(1) \
        .write.option("header",True) \
        .option("delimiter",",") \
        .mode("overwrite") \
        .csv(os.path.join(prata_path, f'prata_fato_TRMM.csv'))

# Salvando no db
def get_single_csv(satelite):
    file_path = os.path.join('2. prata',f'prata_fato_{satelite}.csv',[f for f in os.listdir(os.path.join('2. prata',f'prata_fato_{satelite}.csv')) if f.endswith('.csv')][0])
    return str(Path(file_path).resolve())

satelites = ['AgCFSR','AgMERRA','CHIRPS','CPC','GL','GPM_Final_Run','GPM_Late_Run','PERSIANN_CDR','TRMM']
for satelite_name in satelites:
    conn.execute(f"""
        CREATE OR REPLACE TABLE
                    fato_satelite_{satelite_name} AS 
        SELECT * FROM read_csv_auto(
        '{get_single_csv(satelite_name)}'
        )""")

conn.close()

## fato_estacoes

# Importando os dataframes
estacoes_path = {
    'ANA':r'1. bronze\ana.csv',
    'CEMADEN':r'1. bronze\cemaden.csv',
    'CEMIG':r'1. bronze\cemig.csv',
    'IAC':r'1. bronze\iac.csv',
    'INMET_AGRI':r'1. bronze\inmet_agri.csv',
    'INMET':r'1. bronze\inmet.csv',
    'UNESP':r'1. bronze\unesp.csv'}

estacoes_dict = {
    estacao:pd.read_csv(path) for estacao,path in estacoes_path.items()
}

# Renomeando as colunas
estacoes_dict['ANA'] = estacoes_dict['ANA'] \
    [['id_estacao','data','prec']] \
    .rename(columns={
        'data':'dt_medicao',
        'prec':'vl_precipitacao'
    })

estacoes_dict['CEMADEN'] = estacoes_dict['CEMADEN'] \
    [['id_estacao','data','prec']] \
    .rename(columns={
        'data':'dt_medicao',
        'prec':'vl_precipitacao'
    })

estacoes_dict['CEMIG'] = estacoes_dict['CEMIG'] \
    [['id_estacao','data','tmin','tmed','tmax','prec','urmin','urmax']] \
    .rename(columns={
        'data':'dt_medicao',
        'tmin':'vl_temperatura_minima',
        'tmed':'vl_temperatura_media',
        'tmax':'vl_temperatura_maxima',
        'urmin':'vl_umidade_relativa_minima',
        'urmax':'vl_umidade_relativa_maxima'
    })

estacoes_dict['IAC'] = estacoes_dict['IAC'] \
    [['id_estacao','data','tmin','tmed','tmax','prec']] \
    .rename(columns={
        'data':'dt_medicao',
        'tmin':'vl_temperatura_minima',
        'tmed':'vl_temperatura_media',
        'tmax':'vl_temperatura_maxima'
    })

estacoes_dict['INMET_AGRI'] = estacoes_dict['INMET_AGRI'] \
    [['id_estacao','data','tmin','tmed','tmax','prec','urmin','urmax']] \
    .rename(columns={
        'data':'dt_medicao',
        'tmin':'vl_temperatura_minima',
        'tmed':'vl_temperatura_media',
        'tmax':'vl_temperatura_maxima',
        'urmin':'vl_umidade_relativa_minima',
        'urmax':'vl_umidade_relativa_maxima'
    })

estacoes_dict['INMET'] = estacoes_dict['INMET'] \
    [['id_estacao','data','tmin','tmed','tmax','prec','urmin','urmed','urmax']] \
    .rename(columns={
        'data':'dt_medicao',
        'tmin':'vl_temperatura_minima',
        'tmed':'vl_temperatura_media',
        'tmax':'vl_temperatura_maxima',
        'urmin':'vl_umidade_relativa_minima',
        'urmed':'vl_umidade_relativa_media',
        'urmax':'vl_umidade_relativa_maxima'
    })

estacoes_dict['UNESP'] = estacoes_dict['UNESP'] \
    [['id_estacao','data','tmin','tmed','tmax','prec','urmin','urmax']] \
    .rename(columns={
        'data':'dt_medicao',
        'tmin':'vl_temperatura_minima',
        'tmed':'vl_temperatura_media',
        'tmax':'vl_temperatura_maxima',
        'urmin':'vl_umidade_relativa_minima',
        'urmax':'vl_umidade_relativa_maxima'
    })

for estacao,df, in estacoes_dict.items():
    conn.execute(f"CREATE TABLE IF NOT EXISTS fato_estacao_{estacao} AS SELECT * FROM df")

for estacao,df, in estacoes_dict.items():
    conn.execute(f"DROP TABLE IF EXISTS {estacao}")

conn.execute('SHOW tables').fetch_df()

conn.execute('SELECT * FROM dim_estacoes').fetch_df()

## dim_satelites

table_data = {
    satelite:{
        'dt_min':satelite_dict[satelite].select(F.min('dt_medicao')).collect()[0][0].strftime('%Y-%m-%d'),
        'dt_max':satelite_dict[satelite].select(F.max('dt_medicao')).collect()[0][0].strftime('%Y-%m-%d'),
        'min_lat':satelite_dict[satelite].select(F.min('lat')).collect()[0][0],
        'max_lat':satelite_dict[satelite].select(F.max('lat')).collect()[0][0],
        'delta_lat':compute_lat_delta(satelite_dict[satelite]),
        'min_lon':-satelite_dict[satelite].select(F.min('lon')).collect()[0][0],
        'max_lon':satelite_dict[satelite].select(F.max('lon')).collect()[0][0],
        'delta_lon':compute_lon_delta(satelite_dict[satelite])
    } for satelite in satelite_dict}

dim_satelites = pd.DataFrame(table_data).T.reset_index(names=['nm_satelite'])

conn.execute("CREATE TABLE IF NOT EXISTS dim_satelites AS SELECT * FROM dim_satelites")

conn.close()
