==> 6.Funções_Pipeline.ipynb <==
import duckdb
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os
from pathlib import Path

with open('initial_path.txt','r') as f:
    initial_path = f.read()
    
bronze_path = os.path.join(initial_path,'1. bronze')
prata_path = os.path.join(initial_path,'2. prata')
ouro_path = os.path.join(initial_path,'3. ouro')

conn = duckdb.connect('r2_prata.duckdb')
conn.execute("ATTACH 'r2_ouro.duckdb' AS r2_ouro")


def sql(query,conn=conn):
    return conn.execute(query).fetch_df()
    
show_tables_prata_query = """ SELECT DISTINCT table_name FROM information_schema.tables WHERE table_catalog = 'r2_prata' """
show_tables_ouro_query = """ SELECT DISTINCT table_name FROM information_schema.tables WHERE table_catalog = 'r2_ouro' """

# Funções

## Importar ABT

### import_abt

def import_abt(satelite_list,n_estacoes_vizinhas=4,min_year_training=2010,max_prec_filter=51):
    satelite_columns_str_dict = {
        'AgCFSR':
            """ ,vl_precipitacao_AgCFSR
                ,vl_umidade_relativa_maxima_AgCFSR
                ,vl_radiacao_solar_AgCFSR
                ,vl_temperatura_media_AgCFSR
                ,vl_temperatura_maxima_AgCFSR
                ,vl_temperatura_minima_AgCFSR
                ,vl_vento_velocidade_AgCFSR""",
        'POWER':
            """ ,vl_radiacao_solar_ceu_total_POWER
                ,vl_precipitacao_total_corrigido_POWER
                ,vl_pressao_superficie_POWER
                ,vl_umidade_relativa_2m_POWER
                ,vl_temperatura_maxima_2m_POWER
                ,vl_temperatura_minima_2m_POWER
                ,vl_temperatura_media_2m_POWER
                ,vl_temperatura_orvalho_2m_POWER
                ,vl_direcao_vento_2m_POWER
                ,vl_direcao_vento_10m_POWER
                ,vl_maxima_vento_2m_POWER
                ,vl_ws2m_POWER
                ,vl_maxima_vento_10m_POWER
                ,vl_vento_10m_POWER""",
        'GPM_Late_Run':
            """ ,vl_precipitacao_GPM_Late_Run""",
        'GPM_Final_Run':
            """ ,vl_precipitacao_GPM_Final_Run""",
        'GL':
            """ ,vl_irradiancia_GL""",
        'CPC':
            """ ,vl_precipitacao_CPC""",
        'TRMM':
            """ ,vl_precipitacao_TRMM""",
        'CHIRPS':
            """ ,vl_precipitacao_CHIRPS""",
        'PERSIANN_CDR':
            """ ,vl_precipitacao_PERSIANN_CDR"""
    }

    select_estacoes_vizinhas = '\n'+'\n'.join([f'        ,vl_precipitacao_vizinha_{i+1}' for i in range(n_estacoes_vizinhas)])
    select_satelites = ''.join([satelite_query for satelite,satelite_query in satelite_columns_str_dict.items() if satelite in satelite_list]).replace('    ','  ')
    
    query = f"""
    SELECT
        id_estacao
        ,dt_medicao
        ,MONTH(CAST(dt_medicao AS DATE)) AS dt_mes_medicao
        ,YEAR(CAST(dt_medicao AS DATE)) AS dt_ano_medicao
        ,CASE
            WHEN MONTH(CAST(dt_medicao AS DATE)) IN (11,12,1,2,3) THEN 1
            ELSE 0
        END AS fl_mes_chuvoso
        ,lat_estacao
        ,lon_estacao
        ,vl_temperatura_minima
        ,vl_temperatura_media
        ,vl_temperatura_maxima
        ,vl_umidade_relativa_minima
        ,vl_umidade_relativa_media{select_estacoes_vizinhas}
        ,vl_precipitacao{select_satelites}
    FROM r2_ouro.ABT_satelite_all
    WHERE 1=1
        AND YEAR(CAST(dt_medicao AS DATE)) >= {min_year_training}
        AND vl_precipitacao>=0
        AND vl_precipitacao<={max_prec_filter}
    """

    return sql(query)

## Dados Faltantes

### tratar_dados_faltantes

def tratar_dados_faltantes(df,col,metodo):
    new_df = df.copy()
    if metodo == 'drop':
        new_df = new_df.dropna(subset=[col])
    elif metodo == 'zero':
        new_df[col] = new_df[col].fillna(0)
    elif metodo == 'media_mensal':
        new_df[col] = new_df[col].fillna(new_df.groupby(['dt_mes_medicao','dt_ano_medicao'])[col].transform('mean'))
    elif metodo == 'media_anual':
        new_df[col] = new_df[col].fillna(new_df.groupby(['dt_ano_medicao'])[col].transform('mean'))
    elif metodo == 'media_historico':
        media_historico = new_df['vl_precipitacao'].mean()
        new_df[col] = new_df[col].fillna(media_historico)
    
    return new_df

### print_relatorio_missing

def print_relatorio_missing(df,col,total_rows):
    missing_rows = len(df.loc[df[col].isna()])
    if missing_rows>0:
        print(f'{col} - Missing: {missing_rows}/{total_rows} ({missing_rows*100/total_rows:.2f})')
    return 

### tratar_missing_estacoes_vizinhas

def tratar_missing_estacoes_vizinhas(abt,metodo,print_relatorio=True):

    metodos_possiveis = ['drop','zero','media_mensal','media_anual','media_historico']

    if not metodo in metodos_possiveis:
        raise ValueError(f"Parâmetro '{metodo}' não permitido para o argumento 'metodo'. Utilizar um dos valores possíveis: {', '.join(metodos_possiveis)}.")

    if print_relatorio:
        print("Relatório ESTAÇÕES VIZINHAS: Valores faltantes pré-tratamento")
        total_rows = len(abt)
        for col in abt.columns:
            print_relatorio_missing(abt,col,total_rows)
        print("------------------------------------------------------------------")

    precipitacao_vizinhas_columns = [col for col in abt.columns if 'vl_precipitacao_vizinha' in col]

    new_abt = abt.copy()
    for col in precipitacao_vizinhas_columns:
        new_abt = tratar_dados_faltantes(new_abt,col,metodo)

    if print_relatorio:
        print("Relatório ESTAÇÕES VIZINHAS: Valores faltantes pós-tratamento")
        total_rows = len(new_abt)
        for col in new_abt.columns:
            print_relatorio_missing(new_abt,col,total_rows)
        print("------------------------------------------------------------------")

    return new_abt

### tratar_missing_estacao_base

def tratar_missing_estacao_base(abt,metodo,print_relatorio=True):

    metodos_possiveis = ['drop','zero','media_mensal','media_anual','media_historico']

    if not metodo in metodos_possiveis:
        raise ValueError(f"Parâmetro '{metodo}' não permitido para o argumento 'metodo'. Utilizar um dos valores possíveis: {', '.join(metodos_possiveis)}.")
    
    if print_relatorio:
        print("Relatório ESTAÇÃO BASE: Valores faltantes pré-tratamento")
        total_rows = len(abt)
        for col in abt.columns:
            print_relatorio_missing(abt,col,total_rows)
        print("------------------------------------------------------------------")


    estacao_base_columns = [col for col in abt.columns if col.startswith('vl') and ('temperatura' in col or 'umidade' in col) and not 'POWER' in col]
    
    new_abt = abt.copy()
    for col in estacao_base_columns:
        new_abt = tratar_dados_faltantes(new_abt,col,metodo)

    if print_relatorio:
        print("Relatório ESTAÇÃO BASE: Valores faltantes pós-tratamento")
        total_rows = len(new_abt)
        for col in new_abt.columns:
            print_relatorio_missing(new_abt,col,total_rows)
        print("------------------------------------------------------------------")
            
    return new_abt

abt = import_abt(satelite_list=['CPC','GPM_Final_Run','POWER'],min_year_training=2016)
abt = tratar_missing_estacoes_vizinhas(abt,metodo='media_anual')
abt = tratar_missing_estacao_base(abt,metodo='media_historico')

[col for col in abt.columns if col.startswith('vl') and ('temperatura' in col or 'umidade' in col) and not 'POWER' in col]

abt = tratar_missing_estacoes_vizinhas(abt,metodo='media_anual').dropna()

# Desbalanceamento
